+++ src/Clients/AusterClientsApiClient.php
@@ -50,7 +50,7 @@ class AusterClientsApiClient extends HelaAppClient
     }
 
     /**
-     * @param array<string, mixed> $data
+     * @param array{subscriber_type: 'INDIVIDUAL'|'BUSINESS', numbers: array<int, array{msisdn_ported: string, msisdn_transitory: string, nip: string}>, comments?: string|null} $data
      */
     public function login(array $data): AuthTokenDto
     {
@@ -58,7 +58,7 @@ class AusterClientsApiClient extends HelaAppClient
     }
 
     /**
-     * @param array<string, mixed> $data
+     * @param array{subscriber_type: 'INDIVIDUAL'|'BUSINESS', numbers: array<int, array{msisdn_ported: string, msisdn_transitory: string, nip: string}>, comments?: string|null, auto_submit?: bool} $data
      */
     public function signup(array $data): GenericDto
     {